split convertpathtoapi from nodejsfunction
