+++ .husky/pre-commit
@@ -1,0 +1,1 @@
+npm build
