+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "dist/index.js",
   "scripts": {
-    "build": "tsc"
+    "build": "tsc",
+    "prepare": "husky"
   },
   "author": "Nyman Turkish",
   "license": "ISC",
@@ -14,6 +15,7 @@
   },
   "devDependencies": {
     "@types/node": "^20.11.10",
+    "husky": "^9.1.6",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.3"
   },
